+++ libsafs/global_cached_private.cpp
@@ -968,7 +968,7 @@ public:
 		assert(0);
 	}
 
-	virtual int get_offset_in_first_page() const {
+	virtual off_t get_offset_in_first_page() const {
 		return req->get_offset() % PAGE_SIZE;
 	}
 
@@ -977,7 +977,7 @@ public:
 		return p;
 	}
 
-	virtual int get_size() const {
+	virtual size_t get_size() const {
 		return req->get_size();
 	}
 };
